Move shared code of ArrayFilter.Diff.kt to ArrayFilter.DiffShared.kt
