+++ common/src/main/java/io/github/pelmenstar1/digiDict/common/ArrayFilter.DiffShort.kt
@@ -138,8 +138,8 @@ internal object ArrayFilterDiffShort {
         // allocate forward and backward k-lines. K lines are diagonal lines in the matrix. (see the
         // paper for details)
         // These arrays lines keep the max reachable position for each k-line.
-        val forward = CenteredIntArray(max * 2 + 1)
-        val backward = CenteredIntArray(max * 2 + 1)
+        val forward = ArrayFilterDiffShared.CenteredIntArray(max * 2 + 1)
+        val backward = ArrayFilterDiffShared.CenteredIntArray(max * 2 + 1)
 
         while (stack.size > 0) {
             val range = stack.pop()
